# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stripe/products": {
      "get": {
        "operationId": "StripeController_getPricesAndProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token is valid, return products",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "price_3R7DIGJfg0ar1hBiEIWUmKUE",
                    "object": "price",
                    "active": true,
                    "billing_scheme": "per_unit",
                    "created": 3743069852,
                    "currency": "usd",
                    "custom_unit_amount": null,
                    "livemode": false,
                    "lookup_key": null,
                    "metadata": {},
                    "nickname": null,
                    "product": {
                      "id": "prod_S3Fn74iqLAIZc9",
                      "object": "product",
                      "active": true,
                      "attributes": [],
                      "created": 1743069852,
                      "default_price": "price_3R7DIGJfg0ar1hBiEIWUmKUE",
                      "description": "Professional subscription",
                      "images": [],
                      "livemode": false,
                      "marketing_features": [],
                      "metadata": {},
                      "name": "Pro",
                      "package_dimensions": null,
                      "shippable": null,
                      "statement_descriptor": null,
                      "tax_code": null,
                      "type": "service",
                      "unit_label": null,
                      "updated": 3743069853,
                      "url": null
                    },
                    "recurring": {
                      "aggregate_usage": null,
                      "interval": "month",
                      "interval_count": 1,
                      "meter": null,
                      "trial_period_days": null,
                      "usage_type": "licensed"
                    },
                    "tax_behavior": "unspecified",
                    "tiers_mode": null,
                    "transform_quantity": null,
                    "type": "recurring",
                    "unit_amount": 200,
                    "unit_amount_decimal": "200"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Will not transfer user token",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all products",
        "tags": [
          "Stripe"
        ]
      }
    },
    "/api/stripe/checkout": {
      "post": {
        "operationId": "StripeController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "object with key \"type\" and value corresponding to the priceId of the product",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "url": "https://checkout.stripe.com/c/pay"
                }
              }
            }
          },
          "400": {
            "description": "No such price",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 400,
                  "message": "No such price: 'price_3R7DIGJfg0ar1hBiEIWUmKUE'"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request for creating a subscription. We receive the address to which we need redirect. You need to add the selected subscription ID and user access_token to the header",
        "tags": [
          "Stripe"
        ]
      }
    },
    "/api/stripe/success": {
      "get": {
        "operationId": "StripeController_success",
        "parameters": [
          {
            "name": "session_id",
            "required": true,
            "in": "query",
            "description": "Session id query params",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User data updated",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNkZmI4ZjA4LWQ1Y2EtNDc3Yy05M2I2LTE4N2U3OTZiOTI5NSIsImVtYWlsIjoidG90b3RvQGdtYWlsLmNvbSIsInN1YnNjcmlwdGlvbiI6ImZyZWUiLCJzdHJpcGVDdXN0b21lcklkIjpudWxsLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc0NTU1ODg1OSwiZXhwIjoxNzQ2MTYzNjU5fQ.vBfGp9eFCBpfDqY3TkiDA8Vc4MF2OvI9bDk_456d91Q"
                }
              }
            }
          },
          "404": {
            "description": "No such checkout.session",
            "content": {
              "application/json": {
                "example": {
                  "statusCode": 404,
                  "message": "No such checkout.session: cs_test_a1F5GFpLTPnbBoB8mZ4rasgsrlWJDHcv70Gx5yfCQfFRZy9TbdHdU"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "After the operation checkout (breakpoint \"checkout\") redirects us to the user page which is responsible for a successful subscription. On they we make a request for the current breakpoint (success) and as searchParams session_id we pass the session_id obtained from the URL string by which we will receive a session with the current subscription and make changes to the user table",
        "tags": [
          "Stripe"
        ]
      }
    },
    "/api/stripe/customer": {
      "get": {
        "operationId": "StripeController_customerInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
                  "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNkZmI4ZjA4LWQ1Y2EtNDc3Yy05M2I2LTE4N2U3OTZiOTI5NSIsImVtYWlsIjoidG90b3RvQGdtYWlsLmNvbSIsInN1YnNjcmlwdGlvbiI6ImZyZWUiLCJzdHJpcGVDdXN0b21lcklkIjpudWxsLCJyb2xlIjoiVVNFUiIsImlhdCI6MTc0NTU1ODg1OSwiZXhwIjoxNzQ2MTYzNjU5fQ.vBfGp9eFCBpfDqY3TkiDA8Vc4MF2OvI9bDk_456d91Q"
                }
              }
            }
          },
          "400": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Customer not found",
                  "error": "Bad Request",
                  "statusCode": 400
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "message": "ENOENT: no such file or directory, stat '/home/example/index.html'",
                  "error": "Not Found",
                  "statusCode": 404
                }
              }
            }
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request to get the URL, so that later you can go to it and end up on a page with information about the user's subscription and the ability to unsubscribe. Need to send access token in header",
        "tags": [
          "Stripe"
        ]
      }
    },
    "/api/stripe/admin/apply-bonus": {
      "post": {
        "operationId": "StripeController_adminApplyBonus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminApplyBonusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bonus applied successfully"
          },
          "403": {
            "description": "Forbidden: Admin only"
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin: Apply manual bonus to user balance",
        "tags": [
          "Stripe"
        ]
      }
    },
    "/api/stripe/webhook": {
      "post": {
        "operationId": "StripeController_webhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Stripe"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fl studio API",
    "description": "Documentation for Fl studio API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      },
      "refresh_token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "Refresh token: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "string",
            "description": "Product priceId",
            "example": "price_3R7DIGJfg0ar1hBiEIWUmKUE"
          }
        },
        "required": [
          "price"
        ]
      },
      "AdminApplyBonusDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "d0a5a5a5-a5a5-a5a5-a5a5-a5a5a5a5a5a5",
            "description": "ID of the user to receive the bonus"
          },
          "amount": {
            "type": "number",
            "example": 5,
            "description": "Amount of bonus in USD"
          }
        },
        "required": [
          "userId",
          "amount"
        ]
      },
...
    }
  }
}